+++ flask-data/smarter/resources/info.py
@@ -11,5 +11,14 @@ class SmarterInfoApi(ModelView):
 
     @jwt_required()
     def get(self):
+        """
+        Get information on SMARTER database
+        ---
+        tags:
+          - Info
+        responses:
+          200:
+            description: SMARTER Database status
+        """
         info = self.get_object(pk='smarter')
         return jsonify(info)