+++ HISTORY.md
@@ -8,11 +8,12 @@ TODO
 * support for GIS search in sample endpoints [#13](https://github.com/cnr-ibba/SMARTER-backend/issues/13)
 * configure MongoDB memory usage [#12](https://github.com/cnr-ibba/SMARTER-backend/issues/12)
 * configure logging [#7](https://github.com/cnr-ibba/SMARTER-backend/issues/7)
-* document API [#6](https://github.com/cnr-ibba/SMARTER-backend/issues/6)
 
 0.2.0.dev0
 ----------
 
+* move base location from /api/ to /smarter-api/
+* document API with `flasgger` [#6](https://github.com/cnr-ibba/SMARTER-backend/issues/6)
 * add endpoint for supported assemblies variants [#11](https://github.com/cnr-ibba/SMARTER-backend/issues/11)
 * add info endpoint
 * query with multiple parameters [#23](https://github.com/cnr-ibba/SMARTER-backend/issues/23)